\name{biogeo-package}
\alias{biogeo-package}
\alias{biogeo}
\docType{package}
\title{Point data quality assessment and coordinate conversion}
\description{Functions for error detection and correction in point data quality datasets that are used in species distribution modelling. Includes functions for parsing and converting coordinates into decimal degrees from various formats.
}
\details{
\tabular{ll}{
Package: \tab biogeo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-27\cr
License: \tab LGPL-2\cr
Depends: \tab "R (>= 3.1.0)", "sp (>=1.0-15)", "raster (>= 2.2-31)", "rgdal (>=0.8-16)", "stringr (>= 0.6.2)", "maptools (>= 0.8-30)", "vegan (>=2.0-10)")\cr
}
}
\author{
Mark Robertson

Maintainer: Mark P. Robertson <mrobertson@zoology.up.ac.za>
}
\references{
see tutorial document
}
\keyword{ package }
\seealso{
maptools, raster, dismo}
