% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/climapalette.R
\name{climapalette}
\alias{climapalette}
\title{Climatic Colour Codes for Vegetation Types}
\usage{
climapalette(type, humidity = -PEratio, continentality = Cont,
  temperature = -PET)
}
\arguments{
\item{type}{Community type.}

\item{humidity,continentality,temperature}{Environmental variables
used for red, green and blue in colours.}
}
\description{
Function finds RGB colour codes for vegetation types using three
variables. Choosing climatic variables matching the parameter names
gives maximum values to equatorial rainforest with abundant rains,
equitable non-continental temperature and high temperature.
}
\author{
Jari Oksanen
}

