\name{msk60}
\alias{msk60}
\docType{data}
\title{
msk60
}
\description{
A vector of index numbers indicating the grid cells with environmental data in 1 degree (60 minute) dataset}
\usage{data("msk60")}
\format{
  The format is:
 atomic [1:16216] 2304 2305 2306 2307 2308 ...
 - attr(*, "na.action")=Class 'exclude'  int [1:37784] 1 2 3 4 5 6 7 8 9 10 ...
}
\details{
Generated by reclassifying the Worldclim 10 minute data to 30 minute spatial resolution}
\source{
www.worldclim.org}
\examples{
data(msk60)
}
\keyword{datasets}
