% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PEThorn.R
\name{PEThorn}
\alias{PEThorn}
\title{Thornthwaite's Potential Evapotranspiration}
\usage{
PEThorn(temp, lat, twist = FALSE)
}
\arguments{
\item{temp}{Vector of 12 values of monthly mean temperatures in Centigrade.}

\item{lat}{Latitude in degrees.}

\item{twist}{Adapt model to winter-seasonal climate by adjusting
the equations to the number of months with positive temperature.}
}
\description{
Thornthwaite's Potential Evapotranspiration
}
\references{
Thornthwaite C.W. (1948) An approach toward a rational
classification of climate. \emph{Geographical Review} 38, 55--94.
}

