\name{plotsetup}
\alias{plotsetup}
\title{
Produces devices for plotting}
\description{
Plot windows are used by geo2envid and geo2envpca
}
\usage{
plotsetup(xi, yi)
}
\arguments{
  \item{xi}{
dimension of the x-axis of the plot window in inches}
  \item{yi}{
dimension of the y-axis of the plot window in inches}
}
\value{produces three plot windows}
\author{Mark Robertson}
\seealso{
geo2envid2, geo2envpca}
\examples{
plotsetup(6,6)
}