\name{world}
\alias{world}
\docType{data}
\title{
World country polygons
}
\description{
The dataset comes from the maptools package from wrld_simpl. See data(wrld_simpl) in maptools.
}
\usage{data(world)}
\format{
\describe{see details in maptools}
}
\source{
maptools package
http://mappinghacks.com/data/TM_WORLD_BORDERS_SIMPL-0.2.zip
}
\examples{
data(world)
}
\keyword{datasets}
