% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climapalette.R
\name{climapalette}
\alias{climapalette}
\title{Climatic Colour Codes for Vegetation Types}
\usage{
climapalette(type, humidity, continentality, temperature)
}
\arguments{
\item{type}{Community type.}

\item{humidity, continentality, temperature}{Environmental variables
used for red, green and blue in colours.}
}
\value{
RGB colours for classes plus attributes that give the order
of classes in each colour (\code{rorder}, \code{gorder},
\code{border}). The attributes help in making nicely ordered
legends.
}
\description{
Function finds RGB colour codes for vegetation types using three
variables. Choosing climatic variables matching the parameter names
gives maximum values to equatorial rainforest with abundant rains,
equitable non-continental temperature and high temperature.
}
\examples{
data(climaveg)
clr <- with(climaveg, climapalette(sagetype, -PEratio, Cont, -PET))
with(climaveg, plot(lon, lat, pch=15, cex=0.3, asp=1, col = clr[sagetype]))

}
\author{
Jari Oksanen
}

