% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PEThorn.R
\name{PEThorn}
\alias{PEThorn}
\title{Thornthwaite's Potential Evapotranspiration}
\usage{
PEThorn(temp, lat, monthly = FALSE, twist = FALSE)
}
\arguments{
\item{temp}{Vector of 12 values of monthly mean temperatures in
degrees C.}

\item{lat}{Latitude in degrees.}

\item{monthly}{(logical) Return monthly values instead of annual total.}

\item{twist}{Experimental argument.}
}
\value{
Annual potential evapotranspiration in mm.
}
\description{
Function estimates the Potential Evapotranspiration (PET) from
monthly mean temperatures following Thornthwaite (1948). The
original equations were calibrated for U.S.A., and were the first
attempt at finding PET from readily available meteorological
data. Thornthwaite (1948) used different models for hot and cool
months with limit of 26.5 degrees C. Same monthly mean temperature
generated lower PET in hot than in cool climate, and this was
accommodated for by a \sQuote{heat index}. The cool and hot climate
equations should generate the same (approximately) PET at 26.5
degrees C, but this only happens in the range of \sQuote{heat
index} within U.S.A. excluding Alaska and high altitude Rocky
Mountains. \code{PEThorn} function will restrict the value of the
\code{heat index} into this U.S. range. Details can be interpreted
from the source code.
}
\details{
Function \code{\link[SPEI]{thornthwaite}} (package \pkg{SPEI})
provides an alternative implementation, but it uses the same basic
equations also in hot climates (montly mean > 26.5 degrees), and
does not restrict the range of accepted \sQuote{heat index}. Small
numerical differences are also caused by the algorithms of
estimating the day length.
}
\examples{
PEThorn(c(-9.6,-9.3,-4.8,1.4,7.8,13.5,16.5,14.1,8.9,3.3,-2.8,-7.1),
   lat=65)
}
\author{
Jari Oksanen
}
\references{
Thornthwaite C.W. (1948) An approach toward a rational
classification of climate. \emph{Geographical Review} 38, 55--94.
}
\seealso{
Package \pkg{SPEI} has a wider choice of indices of PET,
among them Hargreaves index and its modification
(\code{\link[SPEI]{hargreaves}}) which is also suitable for
standard meteorological data, and has been calibrated globally.
}

