% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PEThorn.R
\name{PEThorn}
\alias{PEThorn}
\title{Thornthwaite's Potential Evapotranspiration}
\usage{
PEThorn(temp, lat, twist = FALSE)
}
\arguments{
\item{temp}{Vector of 12 values of monthly mean temperatures in Centigrade.}

\item{lat}{Latitude in degrees.}

\item{twist}{Adapt model to winter-seasonal climate by adjusting
the equations to the number of months with positive temperature.}
}
\value{
Annual potential evapotranspiration in mm.
}
\description{
Function estimates the Potential Evapotranspiration (PET) from
monthly mean temperatures following Thornthwaite (1948). The
original equations were calibrated for U.S.A., and were the first
attempt at finding PET from readily available meteorological data.
}
\details{
Function \code{\link[SPEI]{thornthwaite}} (package \pkg{SPEI})
provides an alternative implementation, with two important
differences: (1) Thornthwaite (1948) suggested that the basic
equations should be used only up to 26.5 degrees, and then we
should use to a parabolic model which was given in tabular
form. The \code{PEThorn} uses estimated polynomial coefficients of
this parabola for monthly temperatures over 26.5 degrees, whereas
\pkg{SPEI} package uses the same basic equations which give much
higher values. (2) In winter-seasonal climates, the basic equations
can actually give increased PET with cooler climate and shorter
period of positive temperatures. The \code{PEThorn} function
adjusts the model coefficient (\eqn{I}) for the length of the
period with positive temperatures. This is unpublished, but seems
to improve the estimates in near-arctic conditions.
}
\examples{
PEThorn(c(-9.6,-9.3,-4.8,1.4,7.8,13.5,16.5,14.1,8.9,3.3,-2.8,-7.1),
   lat=65)
}
\author{
Jari Oksanen
}
\references{
Thornthwaite C.W. (1948) An approach toward a rational
classification of climate. \emph{Geographical Review} 38, 55--94.
}
\seealso{
Package \pkg{SPEI} has a wider choice of indices of PET,
among them Hargreaves index and its modification
(\code{\link[SPEI]{hargreaves}}) which is also suitable for
standard meteorological data, and has been calibrated globally.
}

